+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/rule_evaluation_non_decomposable_complete_binned.hpp
@@ -19,9 +19,9 @@ namespace boosting {
     class NonDecomposableCompleteBinnedRuleEvaluationFactory final : public INonDecomposableRuleEvaluationFactory {
         private:
 
-            const float64 l1RegularizationWeight_;
+            const float32 l1RegularizationWeight_;
 
-            const float64 l2RegularizationWeight_;
+            const float32 l2RegularizationWeight_;
 
             const std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr_;
 
@@ -44,7 +44,7 @@ namespace boosting {
              *                                  routines
              */
             NonDecomposableCompleteBinnedRuleEvaluationFactory(
-              float64 l1RegularizationWeight, float64 l2RegularizationWeight,
+              float32 l1RegularizationWeight, float32 l2RegularizationWeight,
               std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr, const Blas& blas, const Lapack& lapack);
 
             std::unique_ptr<IRuleEvaluation<DenseNonDecomposableStatisticVector<float64>>> create(
